Removed all the references about email and username from participant models
