+++ src/Kreta/TaskManager/Domain/Model/Organization/OrganizationParticipant.php
@@ -14,13 +14,10 @@ declare(strict_types=1);
 
 namespace Kreta\TaskManager\Domain\Model\Organization;
 
-use Kreta\SharedKernel\Domain\Model\Identity\EmailAddress;
-use Kreta\SharedKernel\Domain\Model\Identity\Username;
-
 class OrganizationParticipant extends Participant
 {
-    public function __construct(OrganizationParticipantId $id, EmailAddress $email, Username $username)
+    public function __construct(OrganizationParticipantId $id)
     {
-        parent::__construct($id, $email, $username);
+        parent::__construct($id);
     }
 }
