+++ src/Kreta/TaskManager/Domain/Model/Organization/Owner.php
@@ -14,13 +14,10 @@ declare(strict_types=1);
 
 namespace Kreta\TaskManager\Domain\Model\Organization;
 
-use Kreta\SharedKernel\Domain\Model\Identity\EmailAddress;
-use Kreta\SharedKernel\Domain\Model\Identity\Username;
-
 class Owner extends Participant
 {
-    public function __construct(OwnerId $id, EmailAddress $email, Username $username)
+    public function __construct(OwnerId $id)
     {
-        parent::__construct($id, $email, $username);
+        parent::__construct($id);
     }
 }
