+++ src/Kreta/TaskManager/Domain/Model/Organization/Participant.php
@@ -14,22 +14,15 @@ declare(strict_types=1);
 
 namespace Kreta\TaskManager\Domain\Model\Organization;
 
-use Kreta\SharedKernel\Domain\Model\Identity\EmailAddress;
-use Kreta\SharedKernel\Domain\Model\Identity\Username;
-
 abstract class Participant
 {
     protected $id;
     protected $createdOn;
-    protected $email;
-    protected $username;
     protected $updatedOn;
 
-    public function __construct(ParticipantId $id, EmailAddress $email, Username $username)
+    public function __construct(ParticipantId $id)
     {
         $this->id = $id;
-        $this->email = $email;
-        $this->username = $username;
         $this->createdOn = new \DateTimeImmutable();
         $this->updatedOn = new \DateTimeImmutable();
     }
@@ -39,39 +32,7 @@ abstract class Participant
         return $this->id;
     }
 
-    public function changeEmail(EmailAddress $email)
-    {
-        $this->email = $email;
-        $this->updatedOn = new \DateTimeImmutable();
-    }
-
-    public function changeUsername(Username $username)
-    {
-        $this->username = $username;
-        $this->updatedOn = new \DateTimeImmutable();
-    }
-
-    public function createdOn() : \DateTimeInterface
-    {
-        return $this->createdOn;
-    }
-
-    public function email() : EmailAddress
-    {
-        return $this->email;
-    }
-
-    public function updatedOn() : \DateTimeInterface
-    {
-        return $this->updatedOn;
-    }
-
-    public function username() : Username
-    {
-        return $this->username;
-    }
-
-    public function __toString()
+    public function __toString() : string
     {
         return (string) $this->id->id();
     }
