+++ src/Kreta/TaskManager/Domain/Model/User/User.php
@@ -15,19 +15,35 @@ declare(strict_types=1);
 namespace Kreta\TaskManager\Domain\Model\User;
 
 use Kreta\SharedKernel\Domain\Model\AggregateRoot;
+use Kreta\SharedKernel\Domain\Model\Identity\Username;
+use Symfony\Component\Validator\Constraints\Email;
 
 class User extends AggregateRoot
 {
     private $id;
+    private $email;
+    private $username;
 
-    public function __construct(UserId $id)
+    public function __construct(UserId $id, Email $email, Username $username)
     {
         $this->id = $id;
+        $this->email = $email;
+        $this->username = $username;
     }
 
     public function id() : UserId
     {
         return $this->id;
+    }
+
+    public function email() : Email
+    {
+        return $this->email;
+    }
+
+    public function username() : Username
+    {
+        return $this->username;
     }
 
     public function __toString() : string
