+++ src/Kreta/TaskManager/Tests/Spec/Kreta/TaskManager/Application/Organization/CreateOrganizationCommandSpec.php
@@ -20,26 +20,22 @@ class CreateOrganizationCommandSpec extends ObjectBehavior
 {
     function it_can_be_created_with_basic_info()
     {
-        $this->beConstructedWith('user-id', 'Organization name', 'owner@owner.com', 'ownerusername');
+        $this->beConstructedWith('user-id', 'Organization name');
         $this->shouldHaveType(CreateOrganizationCommand::class);
         $this->id()->shouldReturn(null);
         $this->name()->shouldReturn('Organization name');
         $this->slug()->shouldReturn(null);
         $this->userId()->shouldReturn('user-id');
-        $this->ownerEmail()->shouldReturn('owner@owner.com');
-        $this->ownerUsername()->shouldReturn('ownerusername');
     }
 
     function it_can_be_created_with_basic_info_and_organization_id()
     {
-        $this->beConstructedWith('user-id', 'Organization name', 'owner@owner.com', 'ownerusername', 'organization-id');
+        $this->beConstructedWith('user-id', 'Organization name', 'organization-id');
         $this->shouldHaveType(CreateOrganizationCommand::class);
         $this->id()->shouldReturn('organization-id');
         $this->name()->shouldReturn('Organization name');
         $this->slug()->shouldReturn(null);
         $this->userId()->shouldReturn('user-id');
-        $this->ownerEmail()->shouldReturn('owner@owner.com');
-        $this->ownerUsername()->shouldReturn('ownerusername');
     }
 
     function it_can_be_created_with_basic_info_organization_id_owner_id()
@@ -47,8 +43,6 @@ class CreateOrganizationCommandSpec extends ObjectBehavior
         $this->beConstructedWith(
             'user-id',
             'Organization name',
-            'owner@owner.com',
-            'ownerusername',
             'organization-id',
             'owner-id'
         );
@@ -57,8 +51,6 @@ class CreateOrganizationCommandSpec extends ObjectBehavior
         $this->slug()->shouldReturn(null);
         $this->ownerId()->shouldReturn('owner-id');
         $this->userId()->shouldReturn('user-id');
-        $this->ownerEmail()->shouldReturn('owner@owner.com');
-        $this->ownerUsername()->shouldReturn('ownerusername');
     }
 
     function it_can_be_created_with_basic_info_organization_id_owner_id_and_slug()
@@ -66,8 +58,6 @@ class CreateOrganizationCommandSpec extends ObjectBehavior
         $this->beConstructedWith(
             'user-id',
             'Organization name',
-            'owner@owner.com',
-            'ownerusername',
             'organization-id',
             'owner-id',
             'organization-slug'
@@ -77,13 +67,11 @@ class CreateOrganizationCommandSpec extends ObjectBehavior
         $this->slug()->shouldReturn('organization-slug');
         $this->ownerId()->shouldReturn('owner-id');
         $this->userId()->shouldReturn('user-id');
-        $this->ownerEmail()->shouldReturn('owner@owner.com');
-        $this->ownerUsername()->shouldReturn('ownerusername');
     }
 
     function it_cannot_be_created_with_empty_user_id()
     {
-        $this->beConstructedWith('', 'Organization name', 'owner@owner.com', 'ownerusername');
+        $this->beConstructedWith('', 'Organization name');
         $this->shouldThrow(new InvalidArgumentException('User id cannot be null'))->duringInstantiation();
     }
 }
