+++ src/Kreta/TaskManager/Tests/Spec/Kreta/TaskManager/Domain/Model/User/UserSpec.php
@@ -13,16 +13,18 @@
 namespace Spec\Kreta\TaskManager\Domain\Model\User;
 
 use Kreta\SharedKernel\Domain\Model\AggregateRoot;
+use Kreta\SharedKernel\Domain\Model\Identity\Username;
 use Kreta\TaskManager\Domain\Model\User\User;
 use Kreta\TaskManager\Domain\Model\User\UserId;
 use PhpSpec\ObjectBehavior;
+use Symfony\Component\Validator\Constraints\Email;
 
 class UserSpec extends ObjectBehavior
 {
-    function let(UserId $id)
+    function let(UserId $id, Email $email, Username $username)
     {
         $id->id()->willReturn('user-id');
-        $this->beConstructedWith($id);
+        $this->beConstructedWith($id, $email, $username);
     }
 
     function it_can_be_created()
@@ -30,6 +32,8 @@ class UserSpec extends ObjectBehavior
         $this->shouldHaveType(User::class);
         $this->shouldHaveType(AggregateRoot::class);
         $this->id()->shouldReturnAnInstanceOf(UserId::class);
+        $this->email()->shouldReturnAnInstanceOf(Email::class);
+        $this->username()->shouldReturnAnInstanceOf(Username::class);
         $this->__toString()->shouldReturn('user-id');
     }
 }
